qb-back/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
asyncpg
databases
python-dotenv
email-validator
passlib
psycopg2-binary
argon2-cffi
python-jose[cryptography]